chore(weave): extract clickhouse/costs.py from clickhouse_trace_server_batched
Move cost CRUD (create, query, purge) and cost computation helpers
(_get_prices_for_models, _compute_costs_for_buckets) into CostsMixin.
